app/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
openai==1.70.0
pydantic==2.11.1
python-dotenv==1.1.0
